CRM Using ReactJs and Flask/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pyodbc==4.0.39
PyJWT==2.8.0
python-dotenv==1.0.0